requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
mongoengine==0.29.1
psycopg2==2.9.10
pytz==2025.2
sqlparse==0.5.3
tzdata==2025.2
